refactor data validations into service
also group data validation tests together.
