+++ api/src/main/java/au/org/aodn/nrmn/restapi/service/validation/SurveyValidation.java
@@ -1,0 +1,8 @@
+package au.org.aodn.nrmn.restapi.service.validation;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class SurveyValidation {
+    
+}
