+++ api/src/test/java/au/org/aodn/nrmn/restapi/integration/ObservableItemExistsIT.java
@@ -17,7 +17,7 @@ import au.org.aodn.nrmn.restapi.data.model.StagedJob;
 import au.org.aodn.nrmn.restapi.data.model.StagedRow;
 import au.org.aodn.nrmn.restapi.dto.stage.SurveyValidationError;
 import au.org.aodn.nrmn.restapi.enums.ProgramValidation;
-import au.org.aodn.nrmn.restapi.service.validation.ValidationProcess;
+import au.org.aodn.nrmn.restapi.service.validation.DataValidation;
 import au.org.aodn.nrmn.restapi.test.PostgresqlContainerExtension;
 import au.org.aodn.nrmn.restapi.test.annotations.WithTestData;
 
@@ -28,7 +28,7 @@ import au.org.aodn.nrmn.restapi.test.annotations.WithTestData;
 class ObservableItemExistsIT {
 
     @Autowired
-    ValidationProcess validationProcess;
+    DataValidation dataValidation;
 
     @Test
     void notFoundSpeciesShouldFail() {
@@ -37,7 +37,7 @@ class ObservableItemExistsIT {
         StagedRow row = new StagedRow();
         row.setSpecies("Species 20");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, false, Arrays.asList(),
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, false, Arrays.asList(),
                 Arrays.asList(), Arrays.asList(row));
         assertTrue(errors.stream().anyMatch(e -> e.getMessage().equalsIgnoreCase("Species does not exist")));
     }
@@ -49,7 +49,7 @@ class ObservableItemExistsIT {
         StagedRow row = new StagedRow();
         row.setSpecies("Species 56");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, false, Arrays.asList(),
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, false, Arrays.asList(),
                 Arrays.asList(ObservableItem.builder().observableItemName("Species 56").build()), Arrays.asList(row));
         assertFalse(errors.stream().anyMatch(e -> e.getMessage().equalsIgnoreCase("Species does not exist")));
     }
@@ -61,7 +61,7 @@ class ObservableItemExistsIT {
         StagedRow row = new StagedRow();
         row.setSpecies("Survey not done");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, false, Arrays.asList(),
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, false, Arrays.asList(),
                 Arrays.asList(), Arrays.asList(row));
         assertFalse(errors.stream().anyMatch(e -> e.getMessage().equalsIgnoreCase("Species does not exist")));
     }
