+++ api/src/test/java/au/org/aodn/nrmn/restapi/validation/data/ValidationProcessTest.java
@@ -1,4 +1,4 @@
-package au.org.aodn.nrmn.restapi.validation.process;
+package au.org.aodn.nrmn.restapi.validation.data;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -21,7 +21,7 @@ import au.org.aodn.nrmn.restapi.data.repository.ObservationRepository;
 import au.org.aodn.nrmn.restapi.data.repository.SiteRepository;
 import au.org.aodn.nrmn.restapi.dto.stage.SurveyValidationError;
 import au.org.aodn.nrmn.restapi.enums.ProgramValidation;
-import au.org.aodn.nrmn.restapi.service.validation.ValidationProcess;
+import au.org.aodn.nrmn.restapi.service.validation.DataValidation;
 
 @ExtendWith(MockitoExtension.class)
 class ValidationProcessTest {
@@ -36,7 +36,7 @@ class ValidationProcessTest {
     SiteRepository siteRepository;
 
     @InjectMocks
-    ValidationProcess validationProcess;
+    DataValidation dataValidation;
 
     @Test
     void incorrectTimeFormatShouldFail() {
@@ -45,7 +45,7 @@ class ValidationProcessTest {
         StagedRow row = new StagedRow();
         row.setTime("ti:me");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
         assertTrue(errors.stream().anyMatch(e -> e.getMessage().contains("Time format is not valid")));
     }
 
@@ -56,7 +56,7 @@ class ValidationProcessTest {
         StagedRow row = new StagedRow();
         row.setTime("10:15");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
         assertFalse(errors.stream().anyMatch(e -> e.getMessage().contains("Time format is not valid")));
     }
 
@@ -67,7 +67,7 @@ class ValidationProcessTest {
         StagedRow row = new StagedRow();
         row.setTime("40:15");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
         assertTrue(errors.stream().anyMatch(e -> e.getMessage().contains("Time format is not valid")));
     }
 
@@ -92,7 +92,7 @@ class ValidationProcessTest {
         StagedRow row = new StagedRow();
         row.setTime(value);
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
         assertFalse(errors.stream().anyMatch(e -> e.getMessage().contains("Time format is not valid")));
     }
 
@@ -103,7 +103,7 @@ class ValidationProcessTest {
         StagedRow row = new StagedRow();
         row.setTime("");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
         assertFalse(errors.stream().anyMatch(e -> e.getMessage().contains("Time format is not valid")));
     }
 
@@ -114,7 +114,7 @@ class ValidationProcessTest {
         StagedRow row = new StagedRow();
         row.setVis("-99");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
         assertTrue(errors.stream().anyMatch(e -> e.getMessage().contains("Vis is not positive")));
     }
 
@@ -125,7 +125,7 @@ class ValidationProcessTest {
         StagedRow row = new StagedRow();
         row.setVis("");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
         assertFalse(errors.stream().anyMatch(e -> e.getMessage().contains("Vis")));
     }
 
@@ -136,7 +136,7 @@ class ValidationProcessTest {
         StagedRow row = new StagedRow();
         row.setTotal("Not a number");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
         assertTrue(errors.stream().anyMatch(e -> e.getMessage().contains("Total is not an integer")));
     }
 
@@ -147,7 +147,7 @@ class ValidationProcessTest {
         StagedRow row = new StagedRow();
         row.setTotal("10");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
         assertFalse(errors.stream().anyMatch(e -> e.getMessage().contains("Total is not an integer")));
     }
 
@@ -158,7 +158,7 @@ class ValidationProcessTest {
         StagedRow row = new StagedRow();
         row.setLongitude("-67.192519");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
         assertFalse(errors.stream().anyMatch(e -> e.getMessage().contains("Longitude")));
     }
 
@@ -170,7 +170,7 @@ class ValidationProcessTest {
         StagedRow row = new StagedRow();
         row.setDate("not /at /date");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
         assertTrue(errors.stream().anyMatch(e -> e.getMessage().contains("Date format is not valid")));
     }
 
@@ -181,7 +181,7 @@ class ValidationProcessTest {
         StagedRow row = new StagedRow();
         row.setDate("11/09/2018");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
         assertFalse(errors.stream().anyMatch(e -> e.getMessage().contains("Date format is not valid")));
     }
 
@@ -192,7 +192,7 @@ class ValidationProcessTest {
         StagedRow row = new StagedRow();
         row.setDate("11/9/18");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
         assertFalse(errors.stream().anyMatch(e -> e.getMessage().contains("Date format is not valid")));
     }
     
@@ -204,7 +204,7 @@ class ValidationProcessTest {
         StagedRow row = new StagedRow();
         row.setIsInvertSizing("no");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
         assertFalse(errors.stream().anyMatch(e -> e.getMessage().contains("must be 'Yes' or 'No'")));
     }
 
@@ -215,7 +215,7 @@ class ValidationProcessTest {
         StagedRow row = new StagedRow();
         row.setIsInvertSizing("yes");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
         assertFalse(errors.stream().anyMatch(e -> e.getMessage().contains("must be 'Yes' or 'No'")));
     }
 
@@ -227,7 +227,7 @@ class ValidationProcessTest {
         StagedRow row = new StagedRow();
         row.setIsInvertSizing("True");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
         assertTrue(errors.stream().anyMatch(e -> e.getMessage().contains("must be 'Yes' or 'No'")));
     }
 
@@ -239,7 +239,7 @@ class ValidationProcessTest {
         StagedRow row = new StagedRow();
         row.setIsInvertSizing("false");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
         assertTrue(errors.stream().anyMatch(e -> e.getMessage().contains("must be 'Yes' or 'No'")));
     }
 
@@ -251,7 +251,7 @@ class ValidationProcessTest {
         StagedRow row = new StagedRow();
         row.setIsInvertSizing("");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
         assertTrue(errors.stream().anyMatch(e -> e.getMessage().contains("Use Invert Sizing is blank")));
     }
 
@@ -262,7 +262,7 @@ class ValidationProcessTest {
         StagedRow row = new StagedRow();
         row.setIsInvertSizing("");
         row.setStagedJob(job);
-        Collection<SurveyValidationError> errors = validationProcess.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
+        Collection<SurveyValidationError> errors = dataValidation.checkFormatting(ProgramValidation.ATRC, job.getIsExtendedSize(), Arrays.asList(), Arrays.asList(), Arrays.asList(row));
         assertFalse(errors.stream().anyMatch(e -> e.getMessage().contains("must be 'Yes' or 'No'")));
         assertFalse(errors.stream().anyMatch(e -> e.getMessage().contains("Use Invert Sizing is blank")));
     }
